requirements.txt
autopep8==1.5.7
certifi==2021.5.30
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.2.52
Pillow==8.2.0
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.16.0
tifffile==2020.9.3
toml==0.10.2
wincertstore==0.2
